+++ Core/src/iosArm64Main/kotlin/com/sdk/growthbook/Dispatcher.kt
@@ -1,8 +1,0 @@
-package com.sdk.growthbook
-
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-
-// 1
-actual val PlatformDependentIODispatcher: CoroutineDispatcher =
-    Dispatchers.Default